Move types-known (temp until lower-level split)
