+++ packages/types-known/tsconfig.json
@@ -5,5 +5,7 @@
     "outDir": "./build",
     "rootDir": "./src"
   },
-  "references": []
+  "references": [
+    { "path": "../types" }
+  ]
 }
