+++ packages/types/src/known/chain/index.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2021 @polkadot/types-known authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { OverrideVersionedType } from '@polkadot/types/types';
+import type { OverrideVersionedType } from '../../types';
 
 // Type overrides based on specific nodes
 const typesChain: Record<string, OverrideVersionedType[]> = {};
