+++ packages/types/src/known/index.spec.ts
@@ -1,9 +1,8 @@
 // Copyright 2017-2021 @polkadot/types-known authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { TypeRegistry } from '@polkadot/types/create/registry';
-
-import { getModuleTypes } from './';
+import { TypeRegistry } from '../create/registry';
+import { getModuleTypes } from '.';
 
 const registry = new TypeRegistry();
