+++ packages/types/src/known/spec/index.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2021 @polkadot/types-known authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { OverrideVersionedType } from '@polkadot/types/types';
+import type { OverrideVersionedType } from '../../types';
 
 import centrifugeChain from './centrifuge-chain';
 import kusama from './kusama';
