+++ packages/types/src/known/spec/kusama.ts
@@ -3,7 +3,7 @@
 
 /* eslint-disable sort-keys */
 
-import type { OverrideVersionedType } from '@polkadot/types/types';
+import type { OverrideVersionedType } from '../../types';
 
 import { objectSpread } from '@polkadot/util';
