+++ packages/types/src/known/spec/shell.ts
@@ -3,7 +3,7 @@
 
 /* eslint-disable sort-keys */
 
-import type { OverrideVersionedType } from '@polkadot/types/types';
+import type { OverrideVersionedType } from '../../types';
 
 const versioned: OverrideVersionedType[] = [
   {
