+++ packages/types/src/known/upgrades/kusama.ts
