+++ packages/types/src/metadata/v13/toV14.ts
@@ -7,8 +7,9 @@ import type { Text } from '../../primitive/Text';
 import type { Type } from '../../primitive/Type';
 import type { OverrideModuleType, Registry } from '../../types';
 
-import { getModuleTypes, knownOrigins } from '@polkadot/types-known';
 import { stringCamelCase } from '@polkadot/util';
+
+import { getModuleTypes, knownOrigins } from '../../known';
 
 interface MapDef {
   hashers: StorageHasherV13[];
